+++ packages/core/src/transports/connector.test.ts
@@ -28,19 +28,19 @@ test('setup', () => {
   `)
 })
 
-test('behavior: connector type not found', () => {
+test('behavior: connector type not found', async () => {
   const transport = unstable_connector({ type: 'foo' })({})
-  expect(() =>
+  await expect(() =>
     transport.request({ method: 'eth_chainId' }),
   ).rejects.toThrowErrorMatchingInlineSnapshot(`
     [ProviderDisconnectedError: The Provider is disconnected from all chains.
 
     Details: Could not find connector of type "foo" in \`connectors\` passed to \`createConfig\`.
-    Version: viem@2.29.2]
+    Version: viem@2.31.7]
   `)
 })
 
-test('behavior: provider is disconnected', () => {
+test('behavior: provider is disconnected', async () => {
   const transport = unstable_connector(mock)({
     connectors: createStore(() => [
       {
@@ -52,17 +52,17 @@ test('behavior: provider is disconnected', () => {
     ]),
   })
 
-  expect(() =>
+  await expect(() =>
     transport.request({ method: 'eth_chainId' }),
   ).rejects.toThrowErrorMatchingInlineSnapshot(`
     [ProviderDisconnectedError: The Provider is disconnected from all chains.
 
     Details: Provider is disconnected.
-    Version: viem@2.29.2]
+    Version: viem@2.31.7]
   `)
 })
 
-test('behavior: chainId mismatch', () => {
+test('behavior: chainId mismatch', async () => {
   const transport = unstable_connector(mock)({
     chain: optimism,
     connectors: createStore(() => [
@@ -76,22 +76,22 @@ test('behavior: chainId mismatch', () => {
     ]),
   })
 
-  expect(() =>
+  await expect(() =>
     transport.request({ method: 'eth_chainId' }),
   ).rejects.toThrowErrorMatchingInlineSnapshot(`
     [ChainDisconnectedError: The Provider is not connected to the requested chain.
 
     Details: The current chain of the connector (id: 1) does not match the target chain for the request (id: 10 – OP Mainnet).
-    Version: viem@2.29.2]
+    Version: viem@2.31.7]
   `)
 })
 
-test('behavior: request', () => {
+test('behavior: request', async () => {
   const transport = unstable_connector(mock)({
     connectors: createStore(() => [connector]),
   })
 
-  expect(
+  await expect(
     transport.request({ method: 'eth_chainId' }),
   ).resolves.toThrowErrorMatchingInlineSnapshot(`"0x1"`)
 })