+++ packages/react/src/exports/actions/experimental.test.ts
@@ -6,19 +6,19 @@ test('exports', () => {
   expect(Object.keys(experimentalActions)).toMatchInlineSnapshot(`
     [
       "getCallsStatus",
-      "getCapabilities",
-      "sendCalls",
-      "showCallsStatus",
-      "waitForCallsStatus",
-      "writeContracts",
-      "getCallsStatusQueryOptions",
       "getCallsStatusQueryKey",
-      "getCapabilitiesQueryOptions",
+      "getCallsStatusQueryOptions",
+      "getCapabilities",
       "getCapabilitiesQueryKey",
+      "getCapabilitiesQueryOptions",
+      "sendCalls",
       "sendCallsMutationOptions",
+      "showCallsStatus",
       "showCallsStatusMutationOptions",
+      "waitForCallsStatus",
       "waitForCallsStatusQueryKey",
       "waitForCallsStatusQueryOptions",
+      "writeContracts",
       "writeContractsMutationOptions",
     ]
   `)