+++ packages/react/src/hydrate.ts
@@ -1,6 +1,6 @@
 'use client'
 
-import { type ResolvedRegister, type State, hydrate } from '@wagmi/core'
+import { hydrate, type ResolvedRegister, type State } from '@wagmi/core'
 import { type ReactElement, useEffect, useRef } from 'react'
 
 export type HydrateProps = {