+++ site/.vitepress/sidebar.ts
@@ -82,8 +82,12 @@ export function getSidebar() {
             link: '/react/api/connectors',
             items: [
               {
-                text: 'coinbaseWallet',
-                link: '/react/api/connectors/coinbaseWallet',
+                text: 'baseAccount',
+                link: '/react/api/connectors/baseAccount',
+              },
+              {
+                text: 'gemini',
+                link: '/react/api/connectors/gemini',
               },
               { text: 'injected', link: '/react/api/connectors/injected' },
               {
@@ -440,8 +444,12 @@ export function getSidebar() {
             link: '/vue/api/connectors',
             items: [
               {
-                text: 'coinbaseWallet',
-                link: '/vue/api/connectors/coinbaseWallet',
+                text: 'baseAccount',
+                link: '/vue/api/connectors/baseAccount',
+              },
+              {
+                text: 'gemini',
+                link: '/vue/api/connectors/gemini',
               },
               { text: 'injected', link: '/vue/api/connectors/injected' },
               {
@@ -684,8 +692,12 @@ export function getSidebar() {
             link: '/core/api/connectors',
             items: [
               {
-                text: 'coinbaseWallet',
-                link: '/core/api/connectors/coinbaseWallet',
+                text: 'baseAccount',
+                link: '/core/api/connectors/baseAccount',
+              },
+              {
+                text: 'gemini',
+                link: '/core/api/connectors/gemini',
               },
               { text: 'injected', link: '/core/api/connectors/injected' },
               {